DataType Encoding Refactor (#1408)
## Change summary

### `UaDataType`
renamed `UaDataType::getTypeName` to `UaDataType::getEncodingName`

### `UaDecoder` and implementations
signature of struct decoding methods now return `UaStructuredType` instead of `Object`

### `UaEncoder` and implementations
signature of struct encoding methods now requires `UaStructuredType` instead of `Object`

### `DataTypeEncoding` and implementations
modified encode and decode signatures to:
```java
  ExtensionObject encode(EncodingContext context, UaStructuredType struct, NodeId encodingId);

  UaStructuredType decode(EncodingContext context, ExtensionObject encoded, NodeId encodingId);
```

### `DataTypeCodec`
- signature of `decode` now returns `UaStructuredType` instead of `Object`
- signature of `encode` now requires `UaStructuredType` instead of `Object`


### `ExtensionObject`
- decode methods return `UaStructuredType` instead of `Object`
- encode methods require `UaStructuredType` instead of `Object`

### Legacy DataType / DataTypeDictionary
introduces `BsdStructWrapper` as a wrapper type that implements `UaStructuredType`. It's now the type returned by `AbstractBsdCodec::decode` and expected by `AbstractBsdCodec::encode`.

the end result is that the object you get from `ExtensionObject::decode` is now a `BsdStructWrapper` instead of the raw type when using legacy datatypes. The object that should be passed to `ExtensionObject.encode` should be a `BsdStructWrapper` instead of the raw type.

this matches the behavior of the new dynamic DataType support, which returns "wrapper" objects of `DynamicType` or `JsonStruct`.